refactor(web): extract HtmlViewer's deploy/publish flow logic into the file-viewer slice
Cluster E of the FileViewer.tsx decomposition: all deploy/publish state, the
provider-load/config/zone helpers, the deploy-fetch effect, every action
(openDeployModal/openSocialShareFlow/changeDeployProvider/saveDeployConfig/
deployToSelectedProvider/retryDeploymentLink/loadCloudflareZones), and the
derived labels/cards move into hooks/useDeployFlow.hooks.ts, wired into the
orchestrator via one useWiredDeployFlow(...) call. New DeployTransportPort
bundles the deploy/social-share transport; CloudflarePagesZoneOption/
DeployResultCard move to types.ts; resolveShareUrl moves to rules.ts with
origin injected via a new WindowOpenPort.getLocationOrigin(); the modal's
Escape dismiss now goes through the existing DismissPort instead of a bare
window listener (updated the one test that dispatched on window to dispatch
on document, matching the sibling modal's already-migrated Escape test).

The deploy modal + share menu JSX stays inline in the orchestrator for now,
sourcing its values from the hook's return instead of local state/functions;
moving that JSX into dumb components is the next pass on this cluster.
