+++ apps/web/src/components/FileViewer.tsx
@@ -3,23 +3,15 @@ import { createPortal, flushSync } from 'react-dom';
 import { Button, Input, Select } from '@open-design/components';
 import { APP_CHROME_FILE_ACTIONS_ID, APP_CHROME_FILE_ACTIONS_SELECTOR } from './AppChromeHeader';
 import {
-  buildSocialSharePayload,
-  OPEN_DESIGN_GITHUB_REPO_URL,
-  type SocialShareRequest,
-  type SocialShareResponse,
-} from '@open-design/contracts';
-import {
   anonymizeArtifactId,
   artifactKindToTracking,
   type TrackingProjectKind,
-  type TrackingDeployProvider,
 } from '@open-design/contracts/analytics';
 import { useAnalytics } from '../analytics/provider';
 import { exportErrorCode } from '../analytics/export-error-code';
 import { trackIframeLoad } from '../observability/iframe-error';
 import {
   trackArtifactExportResult,
-  trackArtifactDeployResult,
   trackPageView,
   trackShareOptionPopoverClick,
 } from '../analytics/events';
@@ -28,26 +20,13 @@ import { MarkdownRenderer, artifactRendererRegistry } from '../artifacts/rendere
 import { useT, useI18n } from '../i18n';
 import type { Dict, Locale } from '../i18n/types';
 import {
-  checkDeploymentLink,
   CLOUDFLARE_PAGES_PROVIDER_ID,
-  createSocialSharePayload,
-  DEFAULT_DEPLOY_PROVIDER_ID,
-  deployProjectFile,
-  fetchCloudflarePagesZones,
-  fetchDeployConfig,
-  fetchProjectDeployments,
   fetchProjectFiles,
   fetchProjectFileText,
   fetchProjectFileTextPreview,
   uploadProjectFiles,
   projectRawUrl,
-  updateDeployConfig,
-  type WebDeployConfigResponse,
-  type WebCloudflarePagesDeploySelection,
-  type WebDeploymentInfo,
-  type WebDeployProjectFileResponse,
   type WebDeployProviderId,
-  type WebUpdateDeployConfigRequest,
   writeProjectTextFile,
   writeProjectTextFileDetailed,
 } from '../providers/registry';
@@ -162,15 +141,8 @@ import {
   manualEditPreviewShellStyle,
   manualEditFloatingPanelStyle,
   manualEditHoverIconStyle,
-  getDeployProviderOption,
-  normalizeCloudflareDomainPrefixInput,
   isValidCloudflareDomainPrefixInput,
   deployResultState,
-  publicShareUrlForDeployment,
-  deploymentTimestamp,
-  compareDeploymentsByNewest,
-  shareUrlForDeployment,
-  pickLatestShareDeployment,
   mergeManualEditInspectorStyles,
   manualEditInspectorStyleValue,
   manualEditPersistedValueMatchesSavedSnapshot,
@@ -231,6 +203,7 @@ import {
   useWiredViewerToolbarMenus,
   ViewerToolbar,
   MarkdownViewer,
+  useWiredDeployFlow,
 } from '../features/file-viewer';
 import type {
   InspectOverrideMap,
@@ -285,19 +258,6 @@ const IMAGE_EXPORT_FORMAT_OPTIONS: Array<{
   { value: 'jpeg', label: 'JPEG', extension: '.jpg' },
   { value: 'webp', label: 'WebP', extension: '.webp' },
 ];
-type CloudflarePagesZoneOption = {
-  id: string;
-  name: string;
-  status?: string;
-  type?: string;
-};
-type DeployResultCard = {
-  id: string;
-  label: string;
-  url: string;
-  status: string;
-  message?: string;
-};
 // Powered-preview iframe attributes. `allow-same-origin` is what makes real
 // Workers / Web Storage / SharedArrayBuffer possible; it is safe here because
 // the powered iframe loads from the daemon's preview-only loopback host, which
@@ -326,14 +286,6 @@ const HOVER_CARD_DISMISS_DELAY_MS = 80;
 
 // PreviewViewportControls and FileVersionViewportControls now live in the
 // file-viewer slice; imported from the slice barrel above (ADR 0002).
-
-function resolveShareUrl(rawUrl: string): string {
-  const trimmed = rawUrl.trim();
-  if (!trimmed) return '';
-  if (/^https?:\/\//i.test(trimmed)) return trimmed;
-  if (typeof window === 'undefined') return trimmed;
-  return new URL(trimmed, window.location.origin).toString();
-}
 
 // usePreviewCanvasSize now lives in the file-viewer slice as
 // useWiredPreviewCanvasSize (its ResizeObserver/scroll/resize wiring is a
@@ -667,44 +619,7 @@ function HtmlViewer({
   useEffect(() => {
     setPreviewViewportState(getCachedPreviewViewport(fileViewportKey) ?? 'desktop');
   }, [fileViewportKey]);
-  const [deployment, setDeployment] = useState<WebDeploymentInfo | null>(null);
-  const [deploymentsByProvider, setDeploymentsByProvider] = useState<Partial<Record<WebDeployProviderId, WebDeploymentInfo>>>({});
-  const [deployModalOpen, setDeployModalOpen] = useState(false);
-  const [deployModalIntent, setDeployModalIntent] = useState<'deploy' | 'social-share'>('deploy');
-  const closeDeployModal = useCallback(() => {
-    setDeployModalOpen(false);
-    setDeployModalIntent('deploy');
-  }, []);
-  const [deployConfig, setDeployConfig] = useState<WebDeployConfigResponse | null>(null);
-  const [deploying, setDeploying] = useState(false);
-  const [deployPhase, setDeployPhase] = useState<'idle' | 'deploying' | 'preparing-link'>('idle');
-  const [savingDeployConfig, setSavingDeployConfig] = useState(false);
-  const [deployError, setDeployError] = useState<string | null>(null);
-  const [deployResult, setDeployResult] = useState<WebDeployProjectFileResponse | null>(null);
   const deployLinkCopy = useWiredDeployLinkCopy(t);
-  const [deployProviderId, setDeployProviderId] = useState<WebDeployProviderId>(DEFAULT_DEPLOY_PROVIDER_ID);
-  const [projectSocialShare, setProjectSocialShare] = useState<SocialShareResponse | null>(null);
-  const [deployToken, setDeployToken] = useState('');
-  const [teamId, setTeamId] = useState('');
-  const [teamSlug, setTeamSlug] = useState('');
-  const [cloudflareAccountId, setCloudflareAccountId] = useState('');
-  const [cloudflareZones, setCloudflareZones] = useState<CloudflarePagesZoneOption[]>([]);
-  const [cloudflareZonesLoading, setCloudflareZonesLoading] = useState(false);
-  const [cloudflareZonesError, setCloudflareZonesError] = useState<string | null>(null);
-  const [cloudflareZoneId, setCloudflareZoneId] = useState('');
-  const [cloudflareDomainPrefix, setCloudflareDomainPrefix] = useState('');
-  const deployProviderLoadSeqRef = useRef(0);
-  const deployTokenInputRef = useRef<HTMLInputElement | null>(null);
-  useEffect(() => {
-    if (!deployModalOpen) return;
-    const onKeyDown = (event: KeyboardEvent) => {
-      if (event.key !== 'Escape') return;
-      event.preventDefault();
-      closeDeployModal();
-    };
-    window.addEventListener('keydown', onKeyDown);
-    return () => window.removeEventListener('keydown', onKeyDown);
-  }, [closeDeployModal, deployModalOpen]);
   const [inTabPresent, setInTabPresent] = useState(false);
   const [reloadKey, setReloadKey] = useState(0);
   // Set to true permanently once `source` has been populated for the first
@@ -1048,8 +963,6 @@ function HtmlViewer({
     };
   }, [boardImages]);
   const [commentSavedToast, setCommentSavedToast] = useState<string | null>(null);
-  const [deploySavedToast, setDeploySavedToast] = useState<{ message: string; details: string } | null>(null);
-  const [deployActionToast, setDeployActionToast] = useState<string | null>(null);
   const [versionRestoredToast, setVersionRestoredToast] = useState<{ id: number; message: string } | null>(null);
   const versionRestoredToastIdRef = useRef(0);
   const [imageExportModalOpen, setImageExportModalOpen] = useState(false);
@@ -1094,7 +1007,6 @@ function HtmlViewer({
     t,
     onCopyFailed: () => setExportToast({ message: t('useEverywhere.copyFailed'), tone: 'error' }),
   });
-  const [shareGuideToast, setShareGuideToast] = useState<string | null>(null);
   const [selectedSideCommentIds, setSelectedSideCommentIds] = useState<Set<string>>(() => new Set());
   const [commentSidePanelCollapsed, setCommentSidePanelCollapsed] = useState(false);
   const [strokePoints, setStrokePoints] = useState<StrokePoint[]>([]);
@@ -1111,122 +1023,6 @@ function HtmlViewer({
     sidePanelCollapsed: commentSidePanelCollapsed,
     viewport: previewViewport,
   });
-
-  function deploymentMapForCurrentFile(items: WebDeploymentInfo[]) {
-    const next: Partial<Record<WebDeployProviderId, WebDeploymentInfo>> = {};
-    for (const option of DEPLOY_PROVIDER_OPTIONS) {
-      const deploymentForProvider = items
-        .filter((item) => item.fileName === file.name && item.providerId === option.id && item.url?.trim())
-        .sort(compareDeploymentsByNewest)[0];
-      if (deploymentForProvider) next[option.id] = deploymentForProvider;
-    }
-    return next;
-  }
-
-  function syncDeployFormFromConfig(
-    providerId: WebDeployProviderId,
-    config: WebDeployConfigResponse | null,
-  ) {
-    const matchingConfig = config?.providerId === providerId ? config : null;
-    setDeployProviderId(providerId);
-    setDeployConfig(matchingConfig);
-    setDeployToken(matchingConfig?.tokenMask || '');
-    setTeamId(matchingConfig?.teamId || '');
-    setTeamSlug(matchingConfig?.teamSlug || '');
-    setCloudflareAccountId(matchingConfig?.accountId || '');
-    setCloudflareZoneId(matchingConfig?.cloudflarePages?.lastZoneId || '');
-    setCloudflareDomainPrefix(matchingConfig?.cloudflarePages?.lastDomainPrefix || '');
-  }
-
-  function cloudflareConfigHintsFromForm() {
-    const zone = cloudflareZones.find((item) => item.id === cloudflareZoneId);
-    const hints = {
-      ...(cloudflareZoneId.trim() ? { lastZoneId: cloudflareZoneId.trim() } : {}),
-      ...((zone?.name || deployConfig?.cloudflarePages?.lastZoneName)
-        ? { lastZoneName: zone?.name || deployConfig?.cloudflarePages?.lastZoneName }
-        : {}),
-      ...(cloudflareDomainPrefix.trim()
-        ? { lastDomainPrefix: normalizeCloudflareDomainPrefixInput(cloudflareDomainPrefix) }
-        : {}),
-    };
-    return Object.keys(hints).length > 0 ? hints : undefined;
-  }
-
-  function buildDeployConfigRequest(providerId: WebDeployProviderId): WebUpdateDeployConfigRequest {
-    const token = deployToken.trim();
-    if (providerId === CLOUDFLARE_PAGES_PROVIDER_ID) {
-      return {
-        providerId,
-        token,
-        accountId: cloudflareAccountId.trim(),
-        cloudflarePages: cloudflareConfigHintsFromForm(),
-      };
-    }
-    return {
-      providerId,
-      token,
-      teamId: teamId.trim(),
-      teamSlug: teamSlug.trim(),
-    };
-  }
-
-  async function loadDeployProvider(
-    providerId: WebDeployProviderId,
-    options?: { fallbackToExisting?: boolean },
-  ) {
-    const requestSeq = ++deployProviderLoadSeqRef.current;
-    setDeployProviderId(providerId);
-    const deployments = await fetchProjectDeployments(projectId);
-    const nextDeploymentsByProvider = deploymentMapForCurrentFile(deployments);
-    const exactDeployment = nextDeploymentsByProvider[providerId] ?? null;
-    const fallbackDeployment = options?.fallbackToExisting
-      ? Object.values(nextDeploymentsByProvider)[0] ?? null
-      : null;
-    const currentDeployment = exactDeployment ?? fallbackDeployment;
-    // Use the explicit providerId for config/form so a fallback deployment from
-    // another provider only fills the existing-URL display, never the form/credentials.
-    const config = await fetchDeployConfig(providerId);
-    if (requestSeq !== deployProviderLoadSeqRef.current) {
-      return { config: null, currentDeployment: null };
-    }
-    syncDeployFormFromConfig(providerId, config);
-    setDeploymentsByProvider(nextDeploymentsByProvider);
-    setDeployment(currentDeployment ?? null);
-    setDeployResult(currentDeployment ?? null);
-    if (providerId === CLOUDFLARE_PAGES_PROVIDER_ID && config?.configured) {
-      void loadCloudflareZones(config, { requestSeq });
-    }
-    return { config, currentDeployment };
-  }
-
-  async function loadCloudflareZones(
-    config: WebDeployConfigResponse | null = deployConfig,
-    options?: { requestSeq?: number },
-  ) {
-    if (!config?.configured || config.providerId !== CLOUDFLARE_PAGES_PROVIDER_ID) return;
-    const requestSeq = options?.requestSeq ?? deployProviderLoadSeqRef.current;
-    setCloudflareZonesLoading(true);
-    setCloudflareZonesError(null);
-    try {
-      const response = await fetchCloudflarePagesZones();
-      if (requestSeq !== deployProviderLoadSeqRef.current) return;
-      const zones = response?.zones ?? [];
-      setCloudflareZones(zones);
-      const hintedZoneId = response?.cloudflarePages?.lastZoneId || config.cloudflarePages?.lastZoneId || '';
-      const nextZoneId = hintedZoneId && zones.some((zone) => zone.id === hintedZoneId)
-        ? hintedZoneId
-        : zones[0]?.id || '';
-      setCloudflareZoneId(nextZoneId);
-      const hintedPrefix = response?.cloudflarePages?.lastDomainPrefix || config.cloudflarePages?.lastDomainPrefix || '';
-      if (hintedPrefix) setCloudflareDomainPrefix(hintedPrefix);
-    } catch (err) {
-      if (requestSeq !== deployProviderLoadSeqRef.current) return;
-      setCloudflareZones([]);
-      setCloudflareZonesError(err instanceof Error ? err.message : t('fileViewer.cloudflareZonesLoadFailed'));
-    } finally {
-      if (requestSeq === deployProviderLoadSeqRef.current) setCloudflareZonesLoading(false);
-    }
-  }
 
   // Slide deck nav state: the iframe posts the active index + total count
   // back to the host every time a slide settles. Host renders prev/next
@@ -1386,25 +1182,6 @@ function HtmlViewer({
     filesRefreshKey,
     shouldDeferPassivePreviewSource,
   ]);
-
-  useEffect(() => {
-    let cancelled = false;
-    setDeployResult(null);
-    setDeployError(null);
-    deployLinkCopy.resetCopiedDeployLink();
-    setDeployPhase('idle');
-    void fetchProjectDeployments(projectId).then((items) => {
-      if (cancelled) return;
-      const nextDeploymentsByProvider = deploymentMapForCurrentFile(items);
-      const current = nextDeploymentsByProvider[deployProviderId] ?? null;
-      setDeploymentsByProvider(nextDeploymentsByProvider);
-      setDeployment(current ?? null);
-      setDeployResult(current ?? null);
-    });
-    return () => {
-      cancelled = true;
-    };
-  }, [projectId, file.name, deployProviderId]);
 
   const routingHtmlSource = source ?? routingSource ?? lastGoodSourceForRoutingRef.current;
   const passiveLargeHtmlPreview = shouldDeferPassivePreviewSource && source === null;
@@ -3250,211 +3027,6 @@ function HtmlViewer({
     });
   }
 
-  async function openDeployModal(
-    nextProviderId: WebDeployProviderId = deployProviderId,
-    intent: 'deploy' | 'social-share' = 'deploy',
-  ) {
-    setDeployMenuOpen(false);
-    setDeployModalOpen(true);
-    setDeployModalIntent(intent);
-    setDeployError(null);
-    setDeployActionToast(null);
-    deployLinkCopy.resetCopiedDeployLink();
-    setDeployPhase('idle');
-    await loadDeployProvider(nextProviderId, { fallbackToExisting: true });
-  }
-
-  async function openSocialShareFlow() {
-    const providerWithDeployment = DEPLOY_PROVIDER_OPTIONS.find(
-      (option) => deploymentsByProvider[option.id]?.url?.trim(),
-    )?.id;
-    await openDeployModal(providerWithDeployment ?? deployProviderId, 'social-share');
-  }
-
-  async function changeDeployProvider(nextProviderId: WebDeployProviderId) {
-    if (nextProviderId === deployProviderId) return;
-    setDeployError(null);
-    setDeployPhase('idle');
-    await loadDeployProvider(nextProviderId);
-  }
-
-  async function saveDeployConfig() {
-    setSavingDeployConfig(true);
-    setDeployError(null);
-    setDeployActionToast(null);
-    try {
-      if (deployProviderId === CLOUDFLARE_PAGES_PROVIDER_ID) {
-        if (!deployToken.trim()) {
-          setDeployActionToast(t('fileViewer.cloudflareApiTokenRequired'));
-          deployTokenInputRef.current?.focus();
-          return null;
-        }
-        if (!cloudflareAccountId.trim()) {
-          throw new Error(t('fileViewer.cloudflareAccountIdRequired'));
-        }
-      }
-      const config = await updateDeployConfig(buildDeployConfigRequest(deployProviderId));
-      if (!config || config.providerId !== deployProviderId) {
-        throw new Error(t('fileViewer.deployProviderConfigSaveFailed', { provider: deployProviderLabel }));
-      }
-      syncDeployFormFromConfig(deployProviderId, config);
-      if (deployProviderId === CLOUDFLARE_PAGES_PROVIDER_ID) {
-        await loadCloudflareZones(config);
-      }
-      return config;
-    } catch (err) {
-      setDeployError(err instanceof Error ? err.message : t('fileViewer.deployProviderConfigSaveFailed', { provider: deployProviderLabel }));
-      return null;
-    } finally {
-      setSavingDeployConfig(false);
-    }
-  }
-
-  function buildCloudflarePagesDeploySelection(): WebCloudflarePagesDeploySelection | undefined {
-    if (deployProviderId !== CLOUDFLARE_PAGES_PROVIDER_ID) return undefined;
-    const prefix = normalizeCloudflareDomainPrefixInput(cloudflareDomainPrefix);
-    if (!prefix) return undefined;
-    if (!isValidCloudflareDomainPrefixInput(prefix)) {
-      throw new Error(t('fileViewer.cloudflareDomainPrefixInvalid'));
-    }
-    const zone = cloudflareZones.find((item) => item.id === cloudflareZoneId);
-    if (!zone) {
-      throw new Error(t('fileViewer.cloudflareZoneRequired'));
-    }
-    return {
-      zoneId: zone.id,
-      zoneName: zone.name,
-      domainPrefix: prefix,
-    };
-  }
-
-  async function deployToSelectedProvider() {
-    setDeploying(true);
-    setDeployPhase('deploying');
-    setDeployError(null);
-    setDeployActionToast(null);
-    deployLinkCopy.resetCopiedDeployLink();
-    // Real-deploy analytics: report success only after the provider actually
-    // accepts the publish, failed on any hard error / missing config. This is
-    // distinct from the share-popover "opened" signal (artifact_export_result).
-    const deployStarted = performance.now();
-    const providerForTracking: TrackingDeployProvider =
-      deployProviderId === CLOUDFLARE_PAGES_PROVIDER_ID ? 'cloudflare_pages' : 'vercel';
-    const firstConfigure = !deployConfig?.configured;
-    let savedNewToken = false;
-    const fireDeployResult = (
-      result: 'success' | 'failed' | 'cancelled',
-      errorCode?: string,
-    ) => {
-      trackArtifactDeployResult(analytics.track, {
-        page_name: 'artifact',
-        area: 'deploy_modal',
-        artifact_id: anonymizeArtifactId({ projectId, fileName: file.name }),
-        artifact_kind: artifactKindToTracking({ fileKind: file.kind ?? null }),
-        provider: providerForTracking,
-        result,
-        saved_new_token: savedNewToken,
-        first_configure: firstConfigure,
-        ...(errorCode ? { error_code: errorCode } : {}),
-        deploy_duration_ms: Math.round(performance.now() - deployStarted),
-        project_id: projectId,
-        project_kind: projectKind,
-      });
-    };
-    try {
-      const cloudflarePagesSelection = buildCloudflarePagesDeploySelection();
-      const typedToken = deployToken.trim();
-      const hasNewToken = typedToken && typedToken !== deployConfig?.tokenMask;
-      savedNewToken = Boolean(hasNewToken);
-      const cloudflareHints = cloudflareConfigHintsFromForm();
-      const cloudflareHintsChanged = deployProviderId === CLOUDFLARE_PAGES_PROVIDER_ID && Boolean(
-        cloudflareHints?.lastZoneId !== deployConfig?.cloudflarePages?.lastZoneId ||
-        cloudflareHints?.lastZoneName !== deployConfig?.cloudflarePages?.lastZoneName ||
-        cloudflareHints?.lastDomainPrefix !== deployConfig?.cloudflarePages?.lastDomainPrefix,
-      );
-      const needsConfigSave =
-        hasNewToken ||
-        teamId.trim() !== (deployConfig?.teamId || '') ||
-        teamSlug.trim() !== (deployConfig?.teamSlug || '') ||
-        cloudflareAccountId.trim() !== (deployConfig?.accountId || '') ||
-        cloudflareHintsChanged ||
-        !deployConfig?.configured;
-      if (needsConfigSave) {
-        const nextConfig = await saveDeployConfig();
-        if (!nextConfig) {
-          // saveDeployConfig bailed (missing/invalid token, e.g. user clicked
-          // Deploy without entering a key) — count as a failed deploy attempt.
-          fireDeployResult('failed', 'CONFIG_REQUIRED');
-          return;
-        }
-        if (!nextConfig?.configured) {
-          const option = getDeployProviderOption(deployProviderId);
-          throw new Error(t(option.tokenRequiredKey, { provider: t(option.labelKey) }));
-        }
-      }
-      setDeployPhase('preparing-link');
-      const next = await deployProjectFile(projectId, file.name, deployProviderId, cloudflarePagesSelection);
-      setDeploymentsByProvider((current) => ({
-        ...current,
-        [next.providerId]: next,
-      }));
-      setDeployment(next);
-      setDeployResult(next);
-      if (deployResultState(next.status) !== 'failed') {
-        fireDeployResult('success');
-        setDeploySavedToast({
-          message: t('fileViewer.deploySuccessToast'),
-          details: t('fileViewer.deploySuccessToastDetails', {
-            provider: deployProviderLabel,
-            url: next.url,
-          }),
-        });
-      } else {
-        fireDeployResult('failed', `STATUS_${next.status ?? 'UNKNOWN'}`);
-      }
-    } catch (err) {
-      const option = getDeployProviderOption(deployProviderId);
-      const message = err instanceof Error
-        ? err.message
-        : t('fileViewer.deployProviderFailed', { provider: t(option.labelKey) });
-      const tokenRequired =
-        message === t(option.tokenRequiredKey, { provider: t(option.labelKey) });
-      if (tokenRequired) {
-        setDeployActionToast(message);
-        deployTokenInputRef.current?.focus();
-      } else {
-        setDeployError(message);
-      }
-      fireDeployResult(
-        'failed',
-        tokenRequired ? 'CONFIG_REQUIRED' : err instanceof Error ? err.name : 'UNKNOWN',
-      );
-    } finally {
-      setDeploying(false);
-      setDeployPhase('idle');
-    }
-  }
-
-  async function retryDeploymentLink() {
-    const current = deployResult || deployment;
-    if (!current?.id) return;
-    setDeployError(null);
-    setDeployPhase('preparing-link');
-    try {
-      const next = await checkDeploymentLink(projectId, current.id);
-      setDeploymentsByProvider((items) => ({
-        ...items,
-        [next.providerId]: next,
-      }));
-      setDeployment(next);
-      setDeployResult(next);
-    } catch (err) {
-      setDeployError(err instanceof Error ? err.message : t('fileViewer.deployFailed'));
-    } finally {
-      setDeployPhase('idle');
-    }
-  }
-
   function presentInThisTab() {
     setPresentMenuOpen(false);
     setMode('preview');
@@ -4271,212 +3843,93 @@ function HtmlViewer({
       clearBoardComposer();
     }
   }, [activePreviewCommentId, boardMode, effectiveDeck, slideState?.active, visibleSideComments]);
-  const activeDeployment = deployResult || deployment;
-  const activeDeployedUrl = activeDeployment?.url?.trim() || '';
-  const activeDeploymentDelayed = activeDeployment?.status === 'link-delayed';
-  const activeDeploymentProtected = activeDeployment?.status === 'protected';
-  const activeCloudflarePages = activeDeployment?.providerId === CLOUDFLARE_PAGES_PROVIDER_ID
-    ? activeDeployment.cloudflarePages
-    : undefined;
-  const activeCloudflareCustomDomain = activeCloudflarePages?.customDomain;
-  const deployProvider = getDeployProviderOption(deployProviderId);
-  const deployProviderLabel = t(deployProvider.labelKey);
-  const selectedCloudflareZone = cloudflareZones.find((zone) => zone.id === cloudflareZoneId) ?? null;
-  const normalizedCloudflarePrefix = normalizeCloudflareDomainPrefixInput(cloudflareDomainPrefix);
-  const cloudflareHostnamePreview =
-    selectedCloudflareZone && normalizedCloudflarePrefix
-      ? `${normalizedCloudflarePrefix}.${selectedCloudflareZone.name}`
-      : '';
-  const deployResultCards: DeployResultCard[] = activeCloudflarePages
-    ? (() => {
-        const cards: DeployResultCard[] = [];
-        const pagesDevUrl = activeCloudflarePages.pagesDev?.url || activeDeployedUrl;
-        if (pagesDevUrl) {
-          cards.push({
-            id: 'pages-dev',
-            label: t('fileViewer.cloudflarePagesDevLinkLabel'),
-            url: pagesDevUrl,
-            status: activeCloudflarePages.pagesDev?.status || activeDeployment?.status || 'link-delayed',
-            message: activeCloudflarePages.pagesDev?.statusMessage,
-          });
-        }
-        if (activeCloudflareCustomDomain?.url) {
-          cards.push({
-            id: 'custom-domain',
-            label: t('fileViewer.cloudflareCustomDomainLinkLabel'),
-            url: activeCloudflareCustomDomain.url,
-            status: activeCloudflareCustomDomain.status,
-            message:
-              activeCloudflareCustomDomain.errorMessage ||
-              activeCloudflareCustomDomain.statusMessage,
-          });
-        }
-        return cards;
-      })()
-    : activeDeployedUrl
-      ? [{
-          id: 'default',
-          label: activeDeploymentProtected
-            ? t('fileViewer.deployLinkProtectedLabel')
-            : activeDeploymentDelayed
-              ? t('fileViewer.deployLinkPreparingLabel')
-              : t('fileViewer.deployResultLabel'),
-          url: activeDeployedUrl,
-          status: activeDeployment?.status || 'ready',
-          message: activeDeploymentProtected
-            ? t('fileViewer.deployLinkProtected')
-            : activeDeploymentDelayed
-              ? t('fileViewer.deployLinkDelayed')
-              : activeDeployment?.statusMessage,
-        }]
-      : [];
-  const deployActionLabelFor = (providerId: WebDeployProviderId) => {
-    const option = getDeployProviderOption(providerId);
-    const label = t(option.labelKey);
-    const hasActiveDeploymentForProvider = Boolean(deploymentsByProvider[providerId]?.url?.trim());
-    return hasActiveDeploymentForProvider
-      ? t('fileViewer.redeployToProvider', { provider: label })
-      : t('fileViewer.deployToProvider', { provider: label });
-  };
-  const deployedEntries = DEPLOY_PROVIDER_OPTIONS
-    .map((option) => deploymentsByProvider[option.id])
-    .filter((item): item is WebDeploymentInfo => Boolean(item?.url?.trim()));
-  const shareableDeploymentUrl =
-    DEPLOY_PROVIDER_OPTIONS.map((option) => deploymentsByProvider[option.id])
-      .map((item) => publicShareUrlForDeployment(item))
-      .find(Boolean) ?? '';
-  const socialShareBlockedDeployment =
-    shareableDeploymentUrl
-      ? null
-      : deployedEntries.find((item) => deployResultState(item.status) === 'protected' && !publicShareUrlForDeployment(item)) ??
-        deployedEntries.find((item) => !publicShareUrlForDeployment(item)) ??
-        null;
-  const socialShareBlockedState = socialShareBlockedDeployment
-    ? deployResultState(socialShareBlockedDeployment.status)
-    : null;
-  const socialShareDisplayUrl =
-    shareableDeploymentUrl || socialShareBlockedDeployment?.url?.trim() || activeDeployedUrl;
-  const socialShareUnavailableMessage =
-    socialShareBlockedState === 'protected'
-      ? t('fileViewer.deployLinkProtected')
-      : socialShareBlockedState === 'delayed'
-        ? t('fileViewer.deployLinkDelayed')
-        : t('socialShare.deployFirst');
-  const projectSocialShareRequest = useMemo<SocialShareRequest | null>(() => {
-    if (!socialShareDisplayUrl) return null;
-    const title = t('socialShare.projectTitle', { title: exportTitle });
-    const text = t('socialShare.projectText', {
-      title: exportTitle,
-      repo: OPEN_DESIGN_GITHUB_REPO_URL,
-    });
-    return {
-      kind: 'project-html',
-      locale,
-      url: socialShareDisplayUrl,
-      title,
-      text,
-      copyText: t('socialShare.projectCopyText', {
-        title: exportTitle,
-        url: socialShareDisplayUrl,
-        repo: OPEN_DESIGN_GITHUB_REPO_URL,
-      }),
-    };
-  }, [exportTitle, locale, socialShareDisplayUrl, t]);
-  const projectSocialShareFallback = useMemo(
-    () => (projectSocialShareRequest ? buildSocialSharePayload(projectSocialShareRequest) : null),
-    [projectSocialShareRequest],
-  );
-  // Gate the async payload load on a stable *content* key, not the memo's
-  // object identity. The request object can take a fresh identity on renders
-  // where its inputs are value-equal (e.g. while deployment polling re-sets
-  // state with a new map reference), and keying the effect on that identity
-  // made `setProjectSocialShare` re-fire every render — an infinite render
-  // loop once a deployment URL is available (#regression: ready-deploy share).
-  const projectSocialShareKey = projectSocialShareRequest
-    ? JSON.stringify(projectSocialShareRequest)
-    : '';
-  useEffect(() => {
-    setProjectSocialShare(null);
-    if (!projectSocialShareRequest) return;
-    let cancelled = false;
-    void createSocialSharePayload(projectSocialShareRequest)
-      .then((payload) => {
-        if (!cancelled) setProjectSocialShare(payload);
-      })
-      .catch(() => {
-        if (!cancelled) setProjectSocialShare(null);
-      });
-    return () => {
-      cancelled = true;
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [projectSocialShareKey]);
-  const activeProjectSocialShare = projectSocialShare ?? projectSocialShareFallback;
-  const socialShareMenuLabel =
-    activeProjectSocialShare
-      ? t('socialShare.projectSection')
-      : socialShareBlockedState === 'protected'
-        ? t('fileViewer.deployLinkProtectedLabel')
-        : socialShareBlockedState === 'delayed'
-        ? t('fileViewer.deployLinkPreparingLabel')
-          : t('socialShare.deployFirst');
-  const deployActionIconFor = (providerId: WebDeployProviderId) => {
-    if (providerId === 'cloudflare-pages') return 'pages-line';
-    return 'upload-cloud-line';
-  };
-  const latestShareDeployment = useMemo(
-    () => pickLatestShareDeployment(deploymentsByProvider),
-    [deploymentsByProvider],
-  );
-  const latestDeployedShareUrl = latestShareDeployment
-    ? shareUrlForDeployment(latestShareDeployment)
-    : '';
-  const latestShareState = latestShareDeployment
-    ? deployResultState(latestShareDeployment.status)
-    : null;
-  const sharePageUrl = useMemo(
-    () => resolveShareUrl(latestDeployedShareUrl),
-    [latestDeployedShareUrl],
-  );
-  const canCopyShareLink = !streaming && Boolean(sharePageUrl);
-  const canOpenSharePage = !streaming && Boolean(sharePageUrl) && latestShareState !== 'delayed';
-  const shareLinkStatusHint =
-    streaming
-      ? t('fileViewer.shareAfterGenerationComplete')
-      : latestShareState === 'delayed'
-      ? t('fileViewer.deployLinkDelayed')
-      : latestShareState === 'protected'
-        ? t('fileViewer.deployLinkProtected')
-        : '';
-  const shareUnavailableHint = streaming
-    ? t('fileViewer.shareAfterGenerationComplete')
-    : t('fileViewer.shareLinkRequiresDeploy');
-  const shareMenuLabel = t('fileViewer.shareLabel');
-  const deployMenuLabel = t('fileViewer.deployModalTitle') || 'Deploy';
-  const isSocialShareDeployModal = deployModalIntent === 'social-share';
-  const deployModalKicker = isSocialShareDeployModal
-    ? t('socialShare.projectSection')
-    : deployProviderLabel;
-  const deployModalTitle = isSocialShareDeployModal
-    ? t('socialShare.publishPageTitle')
-    : t('fileViewer.deployToProvider', { provider: deployProviderLabel });
-  const deployModalSubtitle = isSocialShareDeployModal
-    ? t('socialShare.publishPageSubtitle')
-    : t('fileViewer.deployModalSubtitle');
-  const deployButtonLabel =
-    deployPhase === 'deploying'
-      ? t('fileViewer.deployingToProvider', { provider: deployProviderLabel })
-      : deployPhase === 'preparing-link'
-        ? t('fileViewer.preparingPublicLink')
-        : isSocialShareDeployModal
-          ? t('socialShare.publishPageTitle')
-          : deployMenuLabel;
-  const statusLabelFor = (state: ReturnType<typeof deployResultState>) => {
-    if (state === 'ready') return t('fileViewer.deployLinkReady');
-    if (state === 'protected') return t('fileViewer.deployLinkProtectedLabel');
-    if (state === 'failed') return t('fileViewer.deployLinkFailed');
-    return t('fileViewer.deployLinkPreparingLabel');
-  };
+  const {
+    deployment,
+    deploymentsByProvider,
+    deployModalOpen,
+    deployModalIntent,
+    closeDeployModal,
+    deployConfig,
+    deploying,
+    deployPhase,
+    savingDeployConfig,
+    deployError,
+    deployResult,
+    deployProviderId,
+    deployToken,
+    setDeployToken,
+    teamId,
+    setTeamId,
+    teamSlug,
+    setTeamSlug,
+    cloudflareAccountId,
+    setCloudflareAccountId,
+    cloudflareZones,
+    cloudflareZonesLoading,
+    cloudflareZonesError,
+    cloudflareZoneId,
+    setCloudflareZoneId,
+    cloudflareDomainPrefix,
+    setCloudflareDomainPrefix,
+    deployTokenInputRef,
+    deploySavedToast,
+    setDeploySavedToast,
+    deployActionToast,
+    setDeployActionToast,
+    shareGuideToast,
+    setShareGuideToast,
+    openDeployModal,
+    openSocialShareFlow,
+    changeDeployProvider,
+    saveDeployConfig,
+    deployToSelectedProvider,
+    retryDeploymentLink,
+    loadCloudflareZones,
+    activeDeployment,
+    activeDeployedUrl,
+    activeDeploymentDelayed,
+    activeDeploymentProtected,
+    deployProvider,
+    deployProviderLabel,
+    selectedCloudflareZone,
+    normalizedCloudflarePrefix,
+    cloudflareHostnamePreview,
+    deployResultCards,
+    deployActionLabelFor,
+    deployActionIconFor,
+    shareableDeploymentUrl,
+    socialShareBlockedDeployment,
+    socialShareBlockedState,
+    socialShareDisplayUrl,
+    socialShareUnavailableMessage,
+    activeProjectSocialShare,
+    socialShareMenuLabel,
+    sharePageUrl,
+    canCopyShareLink,
+    canOpenSharePage,
+    shareLinkStatusHint,
+    shareUnavailableHint,
+    shareMenuLabel,
+    deployMenuLabel,
+    isSocialShareDeployModal,
+    deployModalKicker,
+    deployModalTitle,
+    deployModalSubtitle,
+    deployButtonLabel,
+    statusLabelFor,
+  } = useWiredDeployFlow({
+    projectId,
+    projectKind,
+    fileName: file.name,
+    fileKind: file.kind ?? null,
+    exportTitle,
+    locale,
+    streaming,
+    t,
+    analytics,
+    resetCopiedDeployLink: deployLinkCopy.resetCopiedDeployLink,
+    closeDeployMenu: () => setDeployMenuOpen(false),
+  });
   const initialPreviewLoading = source === null && !sourceEverLoadedRef.current;
   const sourceModeLoading = mode === 'source' && source === null;
   const boardAvailable = mode === 'preview' && source !== null;
