+++ apps/web/src/providers/file-viewer/window-open.ts
@@ -5,3 +5,9 @@ export function openInNewTab(url: string): void {
   if (typeof window === 'undefined') return;
   window.open(url, '_blank', 'noopener,noreferrer');
 }
+
+/** Resolves a relative share path to an absolute URL against the current origin. */
+export function getLocationOrigin(): string {
+  if (typeof window === 'undefined') return '';
+  return window.location.origin;
+}
